feat: ICommonSearchData 타입을 shared로 이동
