+++ src/shared/types/index.ts
@@ -1,1 +1,2 @@
 export * from './image';
+export * from './search';
