+++ src/shared/types/search.ts
@@ -1,0 +1,7 @@
+export interface ICommonSearchData<T> {
+  contents: T;
+  currentPageNumber: number;
+  pageSize: number;
+  numberOfElements: number;
+  hasNextPage: boolean;
+}
